fix(bugs): Fix all 9 bugs from BUG_FIXES1_TODO code review (BUG-1 through BUG-9)
- BUG-1: nvs_storage_get_i2s_pins now propagates hard read errors instead of
  always returning ESP_OK; NOT_FOUND remains silent (expected when key absent)
- BUG-2: nvs_storage_add_paired_device fails fast on blob/name write failure
  before touching the device count, preventing phantom pairing list entries
- BUG-3: cmd_parse guards against UB pointer arithmetic on empty input string
- BUG-4: I2S pin parsing uses strtok_r (thread-safe) instead of strtok
- BUG-5: SAMPLE_RATE and I2S pin parsing use cmd_parse_int instead of atoi,
  so garbage input returns an error rather than silently becoming 0
- BUG-6: bt_streaming_info_t byte counters widened from uint32_t to uint64_t
  to prevent overflow after ~6h of continuous streaming; format strings updated
- BUG-7: beep_overlay_fill clamps frames_to_mix to prevent unsigned underflow
  if a race causes frames_generated >= total_frames at cast time
- BUG-8: audio_processor_beep guards the duration_ms * bytes_per_ms product
  against size_t overflow before casting on 32-bit ESP32
- BUG-9: audio_processor_read checks s_audio_ring != NULL before the drop-drain
  path to prevent NULL dereference during deinit

All 9 fixes include targeted host unit tests. 55/55 host tests pass, 0 warnings.

+++ esp_bt_audio_source/components/bt_manager/include/bt_source.h
@@ -104,15 +104,15 @@ typedef enum {
  */
 typedef struct {
     bt_streaming_state_t state;     // Current streaming state
-    uint32_t bytes_sent;            // DEPRECATED: use bytes_requested instead
+    uint64_t bytes_sent;            // DEPRECATED: use bytes_requested instead
     uint32_t packets_sent;          // Packets sent during streaming
     uint32_t packet_errors;         // Packet errors during streaming
     uint32_t stream_duration;       // in milliseconds
     bool paused;                    // Is streaming paused
     /* CODE_REVIEW5 Task 3.1: Split audio vs silence tracking */
-    uint32_t bytes_requested;       // Total bytes A2DP requested
-    uint32_t bytes_produced;        // Actual audio bytes from queue
-    uint32_t bytes_silence;         // Zero-fill bytes (underruns)
+    uint64_t bytes_requested;       // Total bytes A2DP requested
+    uint64_t bytes_produced;        // Actual audio bytes from queue
+    uint64_t bytes_silence;         // Zero-fill bytes (underruns)
     /* CODE_REVIEW5 Task 3.2: Underrun rate tracking */
     uint32_t underrun_count;        // Number of callbacks with underruns
     uint32_t total_callbacks;       // Total A2DP data callbacks
